Adjust project to use the Traditional Hosting Model. Make some other settings changes too.
The "Traditional Hosting Model" is used a lot out in the wild still.
We will adjust it to the .net 6.0 style at the end of the course, so no worries.

Adjust localhost port settings too.

Remove the Nullable flag to match the project with the course.

Test build the project: dotnet build